bot/requirements.txt
python-telegram-bot==13.15
requests==2.28.2
beautifulsoup4==4.12.2
python-dateutil>=2.8.2